Very quick JOGL 2 port of WWJ (from JOGL 1.1.2-pre-2008-05-23). See JOGL2_Porting.html for a more complete list of substitutions made.
This port was done by adding the new JOGL 2 libraries and natives
followed by extensive search & replace to get WWJ to compile
against JOGL 2 with minimal changes.

Further work is required to use new features in JOGL 2 (such
as proper GLProfile handling and numerous utility objects and
wrappers).

GLProfile.getDefault() is used where necessary and numerous
instances of GL have simply been replaced with GL2 (or equivalent).

Consequently this port is unlikely to run if the actual GL profile
available at runtime is < 2, however this is no loss as it looks like
the JOGL1 version of WWJ required GL2 equivalent
as a minimum anyway.